+++ src/renderer/useSettings.ts
@@ -102,6 +102,8 @@ export const getSettings = (): ConfigurationSchema => {
     language: getConfigValue<string>('language'),
     hideEmptyCategories: getConfigValue<boolean>('hideEmptyCategories'),
     hardwareAcceleration: getConfigValue<boolean>('hardwareAcceleration'),
+    recordCurrentRaidEncountersOnly: getConfigValue<boolean>('recordCurrentRaidEncountersOnly'),
+    uploadCurrentRaidEncountersOnly: getConfigValue<boolean>('uploadCurrentRaidEncountersOnly'),
     /* eslint-enable prettier/prettier */
   };